Massive refactoring to use pseudonyms instead of nicknames for contacts.
The invitation and private messaging UIs are currently broken. Some key
rotation bugs were fixed; others may have been created (unit tests
needed). An encoding for private keys was added. Pseudonyms were moved
out of the messaging package and ratings were moved in.

+++ briar-api/src/net/sf/briar/api/messaging/MessagingConstants.java
@@ -52,5 +52,5 @@ public interface MessagingConstants {
 	 * The timestamp of the oldest message in the database is rounded using
 	 * this modulus to avoid revealing the presence of any particular message.
 	 */
-	long RETENTION_MODULUS = 60 * 60 * 1000; // 1 hour
+	int RETENTION_MODULUS = 60 * 60 * 1000; // 1 hour
 }
